+++ contracts/mocks/AccessControlLightMock.sol
@@ -2,9 +2,9 @@
 
 pragma solidity ^0.8.0;
 
-import "../access/AccessControlLight.sol";
+import "../access/AccessControlEnumerable.sol";
 
-contract AccessControlLightMock is AccessControlLight {
+contract AccessControlEnumerableMock is AccessControlEnumerable {
     constructor() {
         _setupRole(DEFAULT_ADMIN_ROLE, _msgSender());
     }
